Remove running list of circulation events from dashboard. (PP-664) (#161)
## Description

- Removes the running list of circulation events from the "All
authorized libraries" dashboard (it was already removed from the
individual library dashboard pages).
- Deprecates the `showCircEventsDownload` configuration setting property
in favor of the feature flag of the same name. This setting can be
removed once the back end Palace Manager stops sending it.
- Refactors some former class components into function components. 

## Motivation and Context

The running list of events was deemed to be not very useful and the
running list makes repeated requests against the Palace Manager back
end.

[Jira PP-664]

## How Has This Been Tested?

- Manual testing in local environment.
- All tests pass locally.
- [CI
tests](https://github.com/ThePalaceProject/circulation-admin/actions/runs/15525366957).

## Checklist:

- N/A - I have updated the documentation accordingly.
- [X] All new and existing tests passed.