Refactor footer and associated terms of service context. (PP-2670) (#162)
## Description

- Simplify the context manager stack by collapsing the terms of service
context into the higher-level context provider.
- Convert the footer from a class to a functional component.
- Abandon legacy context provider style for context consumer hook.
- Added new tests and refit existing tests for new style context hook.

## Motivation and Context

While working on another issue, it became clear that the current context
provider structure would be an impediment to a good implementation, so
this is just stepping back to address some tech in a little corner of
the app.

Any chance to reduce the footprint of the legacy context provider is
progress for the app overall.

## How Has This Been Tested?

- App testing in local development environment using the dev-server.
- All tests pass locally.
- [CI
tests](https://github.com/ThePalaceProject/circulation-admin/actions/runs/15722949611)
pass.

## Checklist:

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.