Refactor ContextProvider app config settings. (#166)
## Description

Refactors the `ContextProvider` component to accept a single application
`ConfigurationSettings` object in the props, rather than each of its
individual members. `ContextProvider`'s Redux `store` prop remains and
is unchanged.

## Motivation and Context

I ran into this while working on PP-2768. It was proving challenging to
keep types in sync between the legacy `ContextProvider` and other uses
of the app configuration settings. There is more work to do within
`ContextProvider`, but this was the relatively larger chunk to get the
change in and get tests working again.

## How Has This Been Tested?

- Manual testing locally with the dev-server.
- All tests pass locally.
- Tests pass [in
CI](https://github.com/ThePalaceProject/circulation-admin/actions/runs/16622556097).

## Checklist:

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.