Add confirmation when disassociating collection libraries (PP-2408) (#169)
## Description

Adds a confirmation dialog when disassociating a library from a
collection.

## Motivation and Context

When a library is disassociated from a collection, all patron loans and
holds are immediately removed when the change is saved. The confirmation
reduces the risk of accidental disassociation by reminding the user of
the repercussions of the action before it is effected.

## How Has This Been Tested?

- Manually tested locally.
- New tests.
- All tests pass locally.
- [CI
tests](https://github.com/ThePalaceProject/circulation-admin/actions/runs/17078864736)
pass.

## Checklist:

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.