Avoid unintended side-effects on non-interactive library disassociations (PP-2408) (#171)
## Description

Use newly-added `isLibraryRemovalPermitted` (instead of `removeLibrary`)
to confirm interactive removals of library associations from
collections.

NB: This PR is ready for review; however, it depends on #170 for
`isLibraryRemovalPermitted`, so this PR should not be merged until that
one is landed.

## Motivation and Context

Confirmation popups were rendering for each associated library when the
`CollectionEditForm` was being cleared. This is because `removeLibrary`
is used both interactively -- when library `Delete` button is clicked --
and non-interactively, when the form is cleared.

[Jira PP-2408]

## How Has This Been Tested?

- Testing in local environment.
- All tests pass locally.
- [CI
tests](https://github.com/ThePalaceProject/circulation-admin/actions/runs/17103881259)
pass.

## Checklist:

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.